requirements.txt
SoundFile
torchaudio
munch
torch
pydub
pyyaml
librosa
nltk
matplotlib
accelerate
transformers
einops
einops-exts
tqdm
typing
typing-extensions
tensorboard
pandas
phonemizer
enspeak
git+https://github.com/resemble-ai/monotonic_align.git